refactor: move post-login redirection logic to `pages/login` file
- Removes post-login redirection logic from `useUser`
- Moves it to `/login`, the only page where it was needed
